+++ test/data.jl
@@ -17,7 +17,7 @@
     @testset "parsing complex matlab file" begin
         data = parse_matlab_file("../test/data/matlab_02.m")
 
-        @test length(data) == 16
+        @test length(data) == 17
         @test length(data["mpc.bus"]) == 3
         @test length(data["mpc.gen"]) == 3
         @test length(data["mpc.branch"]) == 3
@@ -25,6 +25,7 @@
 
         @test length(data["mpc.areas"][1]) == 2
         @test length(data["mpc.branch_limit"][1]) == 2
+        @test length(data["mpc.zones_named"][1]) == 3
 
         @test isa(data["mpc.version"], SubString{String})
         @test isa(data["mpc.baseMVA"], Float64)
@@ -53,9 +54,10 @@
 
         @test func == "matlab_02"
 
-        @test length(columns) == 4
+        @test length(columns) == 5
 
         @test columns["mpc.areas_named"][2] == "refbus"
+        @test columns["mpc.zones_named"][1] == "index"
 
         for (k,v) in columns
             @test haskey(data, k)